Several Changes to previous commits
I figured what is doing the wrong selection on the marker tap, I removed the marker width/height (had 100), looks like it works better without it, but it might still problematic on a area with a lot of Markers, only solution at some point is to zoom in.

Removed the bool use to control if it is open or not.

 I can't figure out how to remove the one to control which marker is selected, could use some help to figure how I can use a stream on this one!