Problem with Marker width/height
Changing Marker width/height to 25, instead of none, this issue should be looked into.
Looks like it has an effect on it's position on the map, moving all markers off position, any ideias why that happens? If it's set to 100, shows on the place if should be, if we reduce it it goes wrong. Any ideias why this happens?